api/requirements.txt
fastapi
uvicorn
kafka-python
pandas
s3fs
pyarrow
google-genai
python-multipart
sqlalchemy
psycopg2-binary